generator, fix raduis cast warning, minor rtor

+++ GameObjectsGenerator.cpp
@@ -29,32 +29,32 @@ bool GameObjectsGenerator::doStep(sg::IGameEngine* field, long long step) {
 
 
 	// we really should use Command pattern here
-	const float defaultRadius = 0.1f;
+	const double defaultRadius = 0.1;
 	const int32_t defaultPoints = 4;
-	const float defaultColorSeed = 0.5f;
-	const float defaultSpeed = -0.0001f;
+	const double defaultColorSeed = 0.5;
+	const double defaultSpeed = -0.0001;
 
 	auto objectInfo = std::make_shared<sg::GameObjectInfo>();
 	objectInfo->modelName = sqr::Model3d::DEFAULT_MODEL;
 
 	double rndRadius = (double)std::rand() / RAND_MAX;
-	objectInfo->radius = defaultRadius + (0.4f - defaultRadius) * rndRadius;
+	objectInfo->radius = (float)(defaultRadius + (0.4 - defaultRadius) * rndRadius);
 	objectInfo->points = (int32_t)(defaultPoints / objectInfo->radius);
 
 	double rndGreen = (double)std::rand() / RAND_MAX;
-	rndGreen = defaultColorSeed + (1.0f - defaultColorSeed) * rndGreen;
+	rndGreen = defaultColorSeed + (1.0 - defaultColorSeed) * rndGreen;
 
 	double rndBlue = (double)std::rand() / RAND_MAX;
-	rndBlue = defaultColorSeed + (1.0f - defaultColorSeed) * rndBlue;
+	rndBlue = defaultColorSeed + (1.0 - defaultColorSeed) * rndBlue;
 
-	objectInfo->color = glm::vec4{0.0f, rndGreen, rndBlue, 1.0f};
-	objectInfo->speed = glm::vec2{0.0f, defaultSpeed / objectInfo->radius};
+	objectInfo->color = glm::vec4{0.0, rndGreen, rndBlue, 1.0};
+	objectInfo->speed = glm::vec2{0.0, defaultSpeed / objectInfo->radius};
 
 	double rndX = (double)std::rand() / RAND_MAX;
-	rndX = 2.0f * rndX - 1.0f;
-	glm::vec3 position{rndX, 1.0f - objectInfo->radius, 0.0f};
+	rndX = 2.0 * rndX - 1.0;
+	glm::vec3 position{rndX, 1.0 - objectInfo->radius, 0.0};
 
-	glm::vec3 scaling{objectInfo->radius, objectInfo->radius, 0.0f};
+	glm::vec3 scaling{objectInfo->radius, objectInfo->radius, 0.0};
 
 	glm::mat4 transform = glm::translate(glm::mat4{}, position) * glm::scale(glm::mat4{}, scaling);
 	auto gameObject = std::make_shared<sg::GameObject>(field->getRenderer(), field->getWorld(), nextObjectId, objectInfo, transform);
